Add billing log to global billing admin
