+++ app/views/admin/shared/_sidenav_billing.html.haml
@@ -28,3 +28,7 @@
       = tab text("reconcile", model: account_type.constantize.model_name.human(count: 2)),
         [route_name.to_sym, FacilityAccount], (sidenav_tab == route_name)
   = render_view_hook("custom_reconcilable_account_types", sidenav_tab: sidenav_tab)
+
+  - if current_facility.cross_facility? && can?(:manage_billing, current_facility)
+    %li.divider
+    = tab text("billing_log_event"), facility_billing_log_events_path(current_facility), (sidenav_tab == "billing_log")
