+++ config/locales/views/admin/en.billing_log_events.yml
@@ -1,6 +1,6 @@
 en:
   views:
     billing_log_events:
-      index:
+      table:
         statement_email: "!Statement! Email"
         review_orders_email: Review Orders Email
